# cooperativaApi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api Cooperativa",
    "description": "This is api Cooperativa",
    "contact": {
      "name": "Rene Eduardo Recalde",
      "email": "renerecalde2@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/"
    }
  ],
  "tags": [
    {
      "name": "cooperativas",
      "description": "Cooprativas de la provincia."
    }
  ],
  "paths": {
    "/cooperativas": {
      "get": {
        "tags": [
          "cooperativas"
        ],
        "summary": "Todas las cooperativas",
        "description": "Returns todas las cooperativas",
        "operationId": "getCooperativas",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cooperativas"
                }
              }
            }
          },
          "404": {
            "description": "Cooperativa not found",
            "content": {}
          }
        }
      },
      "post": {
        "tags": [
          "cooperativas"
        ],
        "summary": "Creates a new cooperativa.",
        "requestBody": {
          "description": "Create a cooperativa",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CooperativaPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cooperativa"
                }
              }
            }
          }
        }
      }
    },
    "/cooperativas/{id}": {
      "get": {
        "tags": [
          "cooperativas"
        ],
        "summary": "Get an cooperativa",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the cooperativa to get"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cooperativa"
                }
              }
            }
          },
          "404": {
            "description": "Cooperativa not found",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": [
          "cooperativas"
        ],
        "summary": "Delete an cooperativa",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the cooperativa to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cooperativa"
                }
              }
            }
          },
          "404": {
            "description": "Cooperativa not found",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Cooperativas": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Cooperativa"
        }
      },
      "Cooperativa": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "integer",
            "format": "int64"
          },
          "RazonSocial": {
            "type": "string"
          },
          "MatriculaNacional": {
            "type": "string"
          }
        }
      },
      "CooperativaPost": {
        "type": "object",
        "properties": {
          "RazonSocial": {
            "type": "string"
          },
          "MatriculaNacional": {
            "type": "string"
          }
        }
      }
    }
  }
}